third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-dtmf-expected.txt

Tests RTCDTMFSender.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.webkitGetUserMedia({audio:true}, gotStream, error); did not throw exception.
PASS Stream generated.
PASS stream.getAudioTracks().length is 1
PASS stream.getVideoTracks().length is 0
PASS pc is connected
PASS dtmfsender = pc.createDTMFSender(null); threw exception TypeError: Failed to execute 'createDTMFSender' on 'RTCPeerConnection': parameter 1 is not of type 'MediaStreamTrack'..
PASS dtmfsender = pc.createDTMFSender(track); threw exception SyntaxError: Failed to execute 'createDTMFSender' on 'RTCPeerConnection': No local stream is available for the track provided..
PASS dtmfsender = pc.createDTMFSender(track); did not throw exception.
PASS dtmfsender.canInsertDTMF is true
PASS dtmfsender.toneBuffer is "1"
PASS ontonechange was called.
PASS event.tone is "1"
PASS ontonechange was called.
PASS event.tone is ""
PASS successfullyParsed is true

TEST COMPLETE
